# docs/api/map.openapi.yaml
openapi: 3.1.0
info:
  title: KApp Map API
  version: 3.6.0
  summary: Schematic campus map and space search for Fundacion Universitaria Konrad Lorenz.
  description: |
    Digital campus map for Fundacion Universitaria Konrad Lorenz (Bogota, Colombia).

    ## What this service is (and is not)

    This map is **not** a geospatial service. There is no basemap, no GPS and no external map
    provider, and a floor is never positioned by coordinates: it is a drawing, described as data
    and drawn by the client, not a set of floor plan images either. What version 3.1 adds is the
    ground a campus stands on - see below.

    ## Version 2: what the campus survey changed

    Version 1 was designed before anybody walked the campus. The survey of the real buildings
    (`docs/map/LEVANTAMIENTO.md`) broke five of its assumptions, and version 2 fixes each:

    | Found on campus | Version 1 | Version 2 |
    | --- | --- | --- |
    | A mezzanine between the first and second floors; an entrance at `P0`; a terrace numbered 5 | Floors addressed by an integer `level` | Floors addressed by a **`code`** (`S1`, `P0`, `P1`, `MEZZ`, `T`); `level` is a decimal that only orders them |
    | Bienestar's wings are west and east, and its doors carry no wing mark | A closed `NORTE`/`SUR`/`CENTRAL` enum, guessed from a `-N`/`-S`/`-C` suffix | **Wings declared per building**, each with the suffix its doors carry, if any |
...

...
    photographs of posted plans) or `VERIFIED` (walked).

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/map/spaces/search:

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: searchSpaces
      summary: Search or list spaces across the campus
      description: |
        **With `q`:** matched against the door code, the name and the aliases, case- and
        accent-insensitively, ordered by descending relevance. It matches whole words, not
        prefixes: `sistemas` finds the lab, `sistem` does not. Internal codes are never matched,
        so a generated code cannot surface through the search.

        **Without `q`:** every space the filters allow, ordered by building, floor and code -
        how a floor is managed rather than how a person finds a room.

        `type` takes one type code; `category` takes a whole category. Unplaced spaces are
        included.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Spaces]
      parameters:
        - name: q
          in: query
          required: false
          description: Words to search for. At least 2 characters.
          schema: { type: string, minLength: 2, maxLength: 100 }
          example: sala de sistemas
        - $ref: '#/components/parameters/CampusQuery'
        - name: type
          in: query
          required: false
          description: A type code from `GET /api/map/space-types`.
          schema: { type: string, minLength: 1, maxLength: 40 }
          example: LAB
        - name: category
          in: query
          required: false
          schema: { $ref: '#/components/schemas/SpaceCategory' }
        - name: buildingCode
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 10 }
          example: A
        - name: wing
          in: query
          required: false
          description: A wing code of the building.
          schema: { type: string, minLength: 1, maxLength: 8 }
          example: S
        - name: floor
          in: query
          required: false
          description: A floor code; meaningful together with `buildingCode`.
          schema: { type: string, minLength: 1, maxLength: 8 }
          example: P3
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: A page of spaces. An empty page is a normal answer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageResponse' }
              examples:
                Room302: { $ref: '#/components/examples/SearchPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/map/spaces:

    post:
      x-roles: [ROLE_ADMIN]
      operationId: createSpace
      summary: Create a space
      description: |
        Adds one space to a floor. Omit `shape` to inventory it without drawing it. A space whose
        shape covers part of another space's is refused with `409`; every other mistake - a shape
        off the floor, with no area or crossing itself, a door off its outline - is a `400`.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Spaces]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpaceRequest' }
            example:
              code: 503-S
              doorCode: 503-S
              wing: S
              name: Aula 503
              typeCode: CLASSROOM
              buildingCode: EC
              floorCode: P5
              shape: [{ x: 120, y: 320 }, { x: 200, y: 320 }, { x: 200, y: 360 }, { x: 120, y: 360 }]
              doors: [{ from: { x: 130, y: 320 }, to: { x: 150, y: 320 } }]
              accessVia: ASC-SUR
              capacity: 40
      responses:
        '201':
          description: The space was created.
          headers:
            Location:
              description: Path of the created space.
              schema: { type: string }
              example: /api/map/spaces/503-S?buildingCode=EC
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Space' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/map/spaces/{code}:

    parameters:
      - $ref: '#/components/parameters/SpaceCodePath'
      - $ref: '#/components/parameters/BuildingCodeDisambiguator'

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getSpaceByCode
      summary: Resolve a code to its space, floor and building
      description: |
        The endpoint the schedule screen calls when a student taps a class: one round trip from
        the code to the space, its floor and a summary of its building.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Spaces]
      responses:
        '200':
          description: The space, its floor and its building.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            The code exists in more than one building and `buildingCode` was not sent. `details`
            names each candidate building.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                timestamp: '2026-09-23T15:04:05Z'
                status: 409
                error: Conflict
                message: Code 302 exists in more than one building. Pass buildingCode.
                path: /api/map/spaces/302
                details:
                  - { field: buildingCode, issue: 'Candidate building: A' }
                  - { field: buildingCode, issue: 'Candidate building: B' }

    put:
      x-roles: [ROLE_ADMIN]
      operationId: updateSpace
      summary: Replace a space
      description: |
        Replaces a space. Moving one - to another cell, floor or building - is done here too.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Spaces]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpaceRequest' }
      responses:
        '200':
          description: The space as stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Space' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

    delete:
      x-roles: [ROLE_ADMIN]
      operationId: deleteSpace
      summary: Delete a space
      description: |
        Refused with `409` while other spaces name it as their `accessVia` - deleting the lift
        three rooms say to take would leave them pointing at nothing.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Spaces]
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          description: When the service produced this error, in UTC.
          type: string
          format: date-time
        status:
          description: The HTTP status, repeated in the body so a logged response is self-contained.
          type: integer
          example: 400
        error:
          description: The status's standard name - `Bad Request`, `Conflict`. For developers.
          type: string
          example: Bad Request
        message:
          description: One sentence a person can act on. This is the line worth surfacing in a UI.
          type: string
        path:
          description: The path that produced the error.
          type: string
        details:
          description: |
            Field-by-field reasons. `field` names what was wrong - in a layout save, with the
            position of the space: `spaces[3].shape` - and `issue` says what about it.
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
    SpaceCategory:
      type: string
      description: |
        The fixed family a space type belongs to. **Clients choose colours and icons by
        category**; the types under it are data the portal adds to.

        | Value | Spanish | Examples |
        | --- | --- | --- |
        | `TEACHING` | Docencia | aulas, laboratorios, salas de cómputo, auditorios, biblioteca |
        | `PUBLIC_SERVICE` | Atención | ventanillas, recepción, consultorios, enfermería |
        | `OFFICE` | Oficina | direcciones, decanaturas, salas de juntas, archivo |
        | `SOCIAL` | Bienestar y social | cafeterías, gimnasio, salas de juego, terrazas |
        | `FACILITIES` | Servicio | baños, cuartos de aseo y de TI, planta eléctrica, parqueadero |
        | `CIRCULATION` | Circulación | ascensores, escaleras, rampas, entradas |
        | `OTHER` | Sin identificar | drawn on a plan, not yet known |
      enum: [TEACHING, PUBLIC_SERVICE, OFFICE, SOCIAL, FACILITIES, CIRCULATION, OTHER]
    Accessibility:
      type: string
      description: |
        Whether a floor or a space is reachable without stairs. `UNKNOWN` is the default and
        **must not be shown as either answer** - most of the campus has not been checked yet.
      enum: [STEP_FREE, STAIRS_ONLY, UNKNOWN]
    FloorStatus:
      type: string
      description: |
        `UNMAPPED` - the floor exists, nothing is drawn. `DRAFT` - drawn from photographs of posted
        plans, right in shape, unconfirmed in use. `VERIFIED` - walked and corrected on site.
      enum: [UNMAPPED, DRAFT, VERIFIED]
...
    Point:
      type: object
      description: |
        A point on a floor's drawing, in the floor's own units. Origin at the top-left corner as
        the plan hangs; `x` to the right, `y` down.
      required: [x, y]
      properties:
        x: { type: integer, minimum: 0, example: 120 }
        y: { type: integer, minimum: 0, example: 48 }
    Door:
      type: object
      description: |
        A way into a space: the stretch of its outline the door takes up, jamb to jamb. Both ends
        lie on the space's `shape`, along one of its edges.
      required: [from, to]
      properties:
        from: { $ref: '#/components/schemas/Point' }
        to: { $ref: '#/components/schemas/Point' }
    Compass:
      type: string
      enum: [NORTH, EAST, SOUTH, WEST]
      description: |
        A direction on the ground. A floor's `top` says which one its drawing's top edge faces, so a
        client can turn the floor north up or draw a compass that points the right way. Turning a
        drawing a quarter turn clockwise brings what was on its left to the top.
      example: EAST
...
    Bounds:
      type: object
      description: The rectangle around a shape, in the floor's units.
      required: [x, y, width, height]
      properties:
        x: { type: integer, example: 80 }
        y: { type: integer, example: 80 }
        width: { type: integer, example: 160 }
        height: { type: integer, example: 80 }
    Corridor:
      type: object
      description: A walkable route across a floor, drawn as a coloured line.
      required: [code, name, color, path]
      properties:
        code: { type: string, maxLength: 30, example: PAS-CENTRAL }
        name: { type: string, maxLength: 60, example: Pasillo central }
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: CSS hex colour, matching what is painted on the walls where the building colour-codes its wings.
          example: '#5B8DEF'
        path:
          type: array
          description: The points the corridor runs through, in walking order. Must stay on the floor.
          minItems: 1
          items: { $ref: '#/components/schemas/Point' }
    Wing:
      type: object
      description: One arm of a building, as that building names it.
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9]{1,8}$', example: S }
        name: { type: string, maxLength: 60, example: Ala sur }
        doorSuffix:
          type: string
          maxLength: 4
          description: What the doors in this wing append to the room number. Absent when they append nothing.
          example: -S
        note:
          type: string
          maxLength: 300
          description: How to get into or across this wing, where that is not obvious.
          example: Se cruza por el P1 o por la terraza.
    Floor:
      type: object
      description: One floor of a building, described as a drawing the client scales and draws.
      required: [code, level, name, width, height]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9]{1,8}$'
          description: Identifier within the building, and the path segment the API uses. `S1` basement, `P0`, `P1`, `MEZZ`, `T` terrace.
          example: MEZZ
        level:
          type: number
          minimum: -5
          maximum: 99
          description: Vertical order only. Decimal so a mezzanine sits between its floors.
          example: 1.5
        name: { type: string, maxLength: 60, example: Mezzanine }
        status:
          $ref: '#/components/schemas/FloorStatus'
        accessibility:
          $ref: '#/components/schemas/Accessibility'
        note:
          type: string
          maxLength: 300
          description: How to get here when it is not obvious.
          example: Solo por la escalera, desde la recepción.
        width:
          type: integer
          minimum: 1
          maximum: 20000
          description: The drawing's width, in its own units. Every point on the floor lies within it.
          example: 1640
        height: { type: integer, minimum: 1, maximum: 20000, example: 590 }
        top:
          $ref: '#/components/schemas/Compass'
          description: The direction on the ground the drawing's top edge faces. Absent until somebody says - the posted plans are not drawn north up.
        outline:
          type: array
          maxItems: 400
          description: The building's walls around this floor, corners in order. Empty until traced.
          items: { $ref: '#/components/schemas/Point' }
        corridors:
          type: array
          items: { $ref: '#/components/schemas/Corridor' }
        version:
          type: integer
          format: int64
          readOnly: true
          description: Bumped by every layout save. Send it back with the next one; ignored when a whole building is written.
          example: 3
...
    BuildingSummary:
      type: object
      description: A building without its floors.
      required: [id, code, name, campus, aliases, wings]
      properties:
        id: { type: string }
        code: { type: string, example: A }
        name: { type: string, example: Bloque A }
        campus: { type: string, example: Sede Principal }
        description: { type: string }
        aliases:
          type: array
          items: { type: string }
        wings:
          type: array
          items: { $ref: '#/components/schemas/Wing' }
        sinuSedes:
          type: array
          description: The names SINU gives this building as a sede. Since 3.6.
          items: { type: string }
...
    Space:
      type: object
      required: [id, code, name, typeCode, buildingId, buildingCode, campus, floorCode, floorLevel,
                 aliases, doors, effectiveAccessibility]
      properties:
        id: { type: string }
        code:
          type: string
          description: Identifier within the building. **Not for display** - see `doorCode`.
          example: '302'
        doorCode:
          type: string
          description: Exactly as printed on the door. Absent when the door has no number. The only code a client shows.
          example: '302'
        baseCode:
          type: string
          description: '`doorCode` without its wing''s suffix. Absent without a door code.'
          example: '302'
        wing:
          type: string
          description: Code of one of the building's wings.
          example: C
        name: { type: string, example: Laboratorio de Sistemas }
        typeCode: { type: string, example: LAB }
        typeName: { type: string, example: Laboratorio }
        category: { $ref: '#/components/schemas/SpaceCategory' }
        buildingId: { type: string }
        buildingCode: { type: string, example: A }
        campus: { type: string, example: Sede Principal }
        floorCode: { type: string, example: P3 }
        floorLevel: { type: number, example: 3 }
        aliases:
          type: array
          items: { type: string }
        shape:
          type: array
          minItems: 3
          description: The room's walls, corners in order, in the floor's units. Absent while the space is inventoried but not drawn.
          items: { $ref: '#/components/schemas/Point' }
          example: [{ x: 280, y: 200 }, { x: 400, y: 200 }, { x: 400, y: 280 }, { x: 280, y: 280 }]
        bounds:
          $ref: '#/components/schemas/Bounds'
          description: The box around `shape`, for lists and search results that only need a position. Absent with `shape`.
        doors:
          type: array
          description: Openings on the room's walls. Empty when none is drawn yet.
          items: { $ref: '#/components/schemas/Door' }
        accessVia:
          type: string
          description: Code of the lift, staircase or entrance that serves this space - "sube por el ascensor central".
          example: ASC-CENTRAL
        accessibility:
          $ref: '#/components/schemas/Accessibility'
          description: The space's own value; absent when it takes the floor's.
        effectiveAccessibility:
          $ref: '#/components/schemas/Accessibility'
          description: The answer to "can I get there without stairs?", own or inherited.
        note:
          type: string
          description: How to get there when `accessVia` is not enough.
          example: Solo por la escalera norte, desde el P5.
        capacity: { type: integer, minimum: 0, example: 25 }
    SpaceDetail:
      description: A space, its floor and a summary of its building, in one round trip.
      allOf:
        - $ref: '#/components/schemas/Space'
        - type: object
          required: [floor, building]
          properties:
            floor: { $ref: '#/components/schemas/Floor' }
            building: { $ref: '#/components/schemas/BuildingSummary' }
...
    SpaceRequest:
      type: object
      required: [code, name, typeCode, buildingCode, floorCode]
      properties:
        code:
          type: string
          maxLength: 20
          pattern: '^[A-Za-z0-9][A-Za-z0-9-]*$'
          description: The door code for a numbered room; any stable identifier otherwise. Never shown.
        doorCode: { type: string, maxLength: 20, description: Omit when the door has no number. }
        wing: { type: string, maxLength: 8 }
        name: { type: string, minLength: 1, maxLength: 120 }
        typeCode: { type: string, maxLength: 40 }
        buildingCode: { type: string, maxLength: 10 }
        floorCode: { type: string, maxLength: 8 }
        aliases:
          type: array
          items: { type: string, minLength: 1, maxLength: 120 }
        shape:
          type: array
          minItems: 3
          maxItems: 200
          description: Omit to inventory without drawing. Inside the floor, not crossing itself, not overlapping another space.
          items: { $ref: '#/components/schemas/Point' }
        doors:
          type: array
          maxItems: 12
          description: Each on one wall of `shape`. Needs a shape.
          items: { $ref: '#/components/schemas/Door' }
        accessVia: { type: string, maxLength: 20 }
        accessibility:
          $ref: '#/components/schemas/Accessibility'
          description: Omit to take the floor's value.
        note: { type: string, maxLength: 300 }
        capacity: { type: integer, minimum: 0 }
...
    PageResponse:
      type: object
      description: A page of spaces. The only paginated collection in this API.
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Space' }
        page: { type: integer, minimum: 0, example: 0 }
        size: { type: integer, minimum: 1, maximum: 100, example: 20 }
        totalElements: { type: integer, minimum: 0, example: 1 }
        totalPages: { type: integer, minimum: 0, example: 1 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: true }
...
  parameters:
...
    CampusQuery:
      name: campus
      in: query
      required: false
      description: Restrict results to one campus (Sede), matched exactly.
      schema: { type: string, minLength: 1, maxLength: 120 }
      example: Sede Principal
...
    SpaceCodePath:
      name: code
      in: path
      required: true
      description: |
        The space's `code`. For a numbered room it is the door code, which is also what the KApp
        schedule service stores for a class.
      schema: { type: string, minLength: 1, maxLength: 20 }
      example: '302'
    BuildingCodeDisambiguator:
      name: buildingCode
      in: query
      required: false
      description: |
        Needed **only** when the same code exists in more than one building; without it an
        ambiguous code is answered with `409` listing the candidates.
      schema: { type: string, minLength: 1, maxLength: 10 }
      example: A
    PageQuery:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema: { type: integer, minimum: 0, default: 0 }
      example: 0
    SizeQuery:
      name: size
      in: query
      required: false
      description: Page size. Capped at 100 - a larger value is a 400, never silently clamped.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      example: 20
  examples:
...
    SearchPage:
      summary: A search for 'sala de sistemas'
      value:
        content:
          - id: 1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
            code: '302'
            doorCode: '302'
            baseCode: '302'
            wing: C
            name: Laboratorio de Sistemas
            typeCode: LAB
            typeName: Laboratorio
            category: TEACHING
            buildingId: 3f6d1c92-8a41-4b0e-9c77-1d2e3f4a5b60
            buildingCode: A
            campus: Sede Principal
            floorCode: P3
            floorLevel: 3
            aliases: [sala de sistemas, lab 302, laboratorio 302]
            shape: [{ x: 280, y: 200 }, { x: 400, y: 200 }, { x: 400, y: 280 }, { x: 280, y: 280 }]
            bounds: { x: 280, y: 200, width: 120, height: 80 }
            doors: [{ from: { x: 280, y: 230 }, to: { x: 280, y: 250 } }]
            accessVia: ASC-CENTRAL
            effectiveAccessibility: STEP_FREE
            capacity: 25
        page: 0
        size: 20
        totalElements: 1
        totalPages: 1
        first: true
        last: true
  responses:
...
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Authenticated but not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflicts with current state
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }